+++ apps/e2e/forms-storybook/src/app/single-file-attachment/single-file-attachment.component.ts
@@ -17,6 +17,7 @@ export class SingleFileAttachmentComponent {
   public imageAttachment: UntypedFormControl;
   public fileAttachment: UntypedFormControl;
   public attachment: UntypedFormControl;
+  public disabled: UntypedFormControl;
 
   public fileForm: UntypedFormGroup;
 
@@ -25,6 +26,10 @@ export class SingleFileAttachmentComponent {
   }
 
   constructor(formBuilder: UntypedFormBuilder) {
+    this.disabled = new UntypedFormControl({
+      value: undefined,
+      disabled: true,
+    });
     this.attachment = new UntypedFormControl(undefined, Validators.required);
     this.imageAttachment = new UntypedFormControl(
       {