+++ libs/components/forms/src/lib/modules/field-group/field-group.component.ts
@@ -6,6 +6,9 @@ import {
   booleanAttribute,
   numberAttribute,
 } from '@angular/core';
+import { SkyIdModule } from '@skyux/core';
+
+import { SkyFormFieldLabelTextRequiredService } from '../shared/form-field-label-text-required.service';
 
 import { SkyFieldGroupHeadingLevel } from './field-group-heading-level';
 import { SkyFieldGroupHeadingStyle } from './field-group-heading-style';
@@ -18,7 +21,8 @@ import { SkyFieldGroupHeadingStyle } from './field-group-heading-style';
   templateUrl: './field-group.component.html',
   styleUrl: './field-group.component.scss',
   standalone: true,
-  imports: [CommonModule],
+  imports: [CommonModule, SkyIdModule],
+  providers: [SkyFormFieldLabelTextRequiredService],
 })
 export class SkyFieldGroupComponent {
   /**
@@ -27,6 +31,14 @@ export class SkyFieldGroupComponent {
    */
   @Input({ required: true })
   public labelText!: string;
+
+  /**
+   * [Persistent inline help text](https://developer.blackbaud.com/skyux/design/guidelines/user-assistance#inline-help) that provides
+   * additional context to the user.
+   * @preview
+   */
+  @Input()
+  public hintText: string | undefined;
 
   /**
    * Indicates whether to hide the `labelText`.