+++ libs/components/forms/src/lib/modules/file-attachment/file-attachment.component.spec.ts
@@ -19,6 +19,8 @@ import {
 
 import { BehaviorSubject } from 'rxjs';
 
+import { SkyFormFieldLabelTextRequiredService } from '../shared/form-field-label-text-required.service';
+
 import { SkyFileAttachmentComponent } from './file-attachment.component';
 import { SkyFileItem } from './file-item';
 import { FileAttachmentTestComponent } from './fixtures/file-attachment.component.fixture';
@@ -385,9 +387,11 @@ describe('File attachment', () => {
     const input = getInputDebugEl(fixture);
 
     expect(input.nativeElement.getAttribute('required')).not.toBeNull();
-    expect(labelWrapper?.classList.contains('sky-control-label-required')).toBe(
-      true,
-    );
+    expect(
+      fixture.nativeElement
+        .querySelector('span.sky-control-label')
+        .classList.contains('sky-control-label-required'),
+    ).toBe(true);
     expect(
       labelWrapper?.querySelector('.sky-screen-reader-only')?.textContent,
     ).toBe('Required');
@@ -468,7 +472,6 @@ describe('File attachment', () => {
   }));
 
   it('should handle removing the labelText', fakeAsync(() => {
-    fixture.componentInstance.required = true;
     fixture.componentInstance.labelText = 'label text';
     fixture.componentInstance.labelElementText = undefined;
     fixture.componentInstance.showLabel = false;
@@ -478,18 +481,16 @@ describe('File attachment', () => {
     tick();
     fixture.detectChanges();
 
-    const labelWrapper = getLabelWrapper();
-
-    expect(labelWrapper?.classList.contains('sky-control-label-required')).toBe(
-      true,
-    );
+    expect(
+      fixture.nativeElement.querySelector('span.sky-control-label'),
+    ).toBeDefined();
 
     fixture.componentInstance.labelText = undefined;
     fixture.detectChanges();
 
-    expect(labelWrapper?.classList.contains('sky-control-label-required')).toBe(
-      false,
-    );
+    expect(
+      fixture.nativeElement.querySelector('span.sky-control-label'),
+    ).toBeNull();
   }));
 
   it('should click the file input on choose file button click', () => {
@@ -1517,6 +1518,33 @@ describe('File attachment', () => {
     validateLabelText('label element');
   });
 
+  it('should not render if a parent component requires label text and it is not provided', () => {
+    TestBed.resetTestingModule();
+    TestBed.configureTestingModule({
+      imports: [FileAttachmentTestModule, SkyThemeModule],
+      providers: [
+        {
+          provide: SkyThemeService,
+          useValue: mockThemeSvc,
+        },
+        SkyFormFieldLabelTextRequiredService,
+      ],
+    });
+
+    const fixture = TestBed.createComponent(FileAttachmentTestComponent);
+    const fileAttachment = fixture.nativeElement.querySelector(
+      'sky-file-attachment',
+    );
+    const labelTextRequiredSvc = TestBed.inject(
+      SkyFormFieldLabelTextRequiredService,
+    );
+    const labelTextSpy = spyOn(labelTextRequiredSvc, 'validateLabelText');
+    fixture.detectChanges();
+
+    expect(labelTextSpy).toHaveBeenCalled();
+    expect(fileAttachment).not.toBeVisible();
+  });
+
   it('should mark as dirty when an invalid file is uploaded first', () => {
     const files = [
       {
@@ -1532,6 +1560,54 @@ describe('File attachment', () => {
     fixture.detectChanges();
 
     expect(fixture.componentInstance.attachment.dirty).toBeTrue();
+  });
+
+  it('should render help inline with popover only if label text is provided', () => {
+    fixture.componentInstance.popoverContent = 'popover content';
+    fixture.componentInstance.showLabel = false;
+    fixture.detectChanges();
+
+    expect(
+      fixture.nativeElement.querySelectorAll('sky-help-inline').length,
+    ).toBe(0);
+
+    fixture.componentInstance.labelText = 'labelText';
+    fixture.detectChanges();
+
+    expect(
+      fixture.nativeElement.querySelectorAll('sky-help-inline').length,
+    ).toBe(1);
+  });
+
+  it('should not render help inline for popover unless popover content is set', () => {
+    fixture.componentInstance.popoverTitle = 'popover title';
+    fixture.componentInstance.showLabel = false;
+    fixture.componentInstance.labelText = 'labelText';
+    fixture.detectChanges();
+
+    expect(
+      fixture.nativeElement.querySelectorAll('sky-help-inline').length,
+    ).toBe(0);
+
+    fixture.componentInstance.popoverContent = 'popover content';
+    fixture.detectChanges();
+
+    expect(
+      fixture.nativeElement.querySelectorAll('sky-help-inline').length,
+    ).toBe(1);
+  });
+
+  it('should render hint if `hintText` is set', () => {
+    const hintText = 'hint text';
+    fixture.componentInstance.hintText = hintText;
+    fixture.detectChanges();
+
+    const hintEl = fixture.nativeElement.querySelector(
+      '.sky-file-attachment-hint-text',
+    );
+
+    expect(hintEl).not.toBeNull();
+    expect(hintEl?.textContent.trim()).toBe(hintText);
   });
 });