+++ libs/components/forms/src/lib/modules/file-attachment/file-attachments.module.ts
@@ -2,6 +2,7 @@ import { CommonModule, DecimalPipe } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { SkyIdModule, SkyTrimModule } from '@skyux/core';
+import { SkyHelpInlineModule } from '@skyux/help-inline';
 import { SkyIconModule } from '@skyux/indicators';
 import { SkyThemeModule } from '@skyux/theme';
 
@@ -34,6 +35,7 @@ import { SkyFileSizePipe } from './file-size.pipe';
     SkyIdModule,
     SkyThemeModule,
     SkyTrimModule,
+    SkyHelpInlineModule,
   ],
   exports: [
     SkyFileAttachmentComponent,