+++ libs/components/forms/src/lib/modules/file-attachment/file-drop.component.spec.ts
@@ -4,6 +4,8 @@ import { By } from '@angular/platform-browser';
 import { SkyAppTestUtility, expect, expectAsync } from '@skyux-sdk/testing';
 import { SkyIdService, SkyLiveAnnouncerService } from '@skyux/core';
 
+import { SkyFormFieldLabelTextRequiredService } from '../shared/form-field-label-text-required.service';
+
 import { SkyFileAttachmentsModule } from './file-attachments.module';
 import { SkyFileDropComponent } from './file-drop.component';
 import { SkyFileItem } from './file-item';
@@ -37,6 +39,8 @@ describe('File drop component', () => {
     );
 
     fixture = TestBed.createComponent(SkyFileDropComponent);
+    fixture.detectChanges();
+
     el = fixture.nativeElement;
     componentInstance = fixture.componentInstance;
 
@@ -389,6 +393,26 @@ describe('File drop component', () => {
 
     expect(labelEl).not.toBeNull();
     expect(labelEl).toHaveCssClass('sky-screen-reader-only');
+  });
+
+  it('should not render if a parent component requires label text and it is not provided', () => {
+    TestBed.resetTestingModule();
+
+    TestBed.configureTestingModule({
+      imports: [SkyFileAttachmentsModule],
+      declarations: [FileDropContentComponent],
+      providers: [SkyFormFieldLabelTextRequiredService],
+    });
+
+    const fixture = TestBed.createComponent(SkyFileDropComponent);
+    const labelTextRequiredSvc = TestBed.inject(
+      SkyFormFieldLabelTextRequiredService,
+    );
+    const labelTextSpy = spyOn(labelTextRequiredSvc, 'validateLabelText');
+    fixture.detectChanges();
+
+    expect(labelTextSpy).toHaveBeenCalled();
+    expect(fixture.nativeElement).not.toBeVisible();
   });
 
   it('should render the hintText when provided', () => {
@@ -1243,6 +1267,31 @@ describe('File drop component', () => {
       linkInput.nativeElement.attributes.getNamedItem('aria-describedby').value,
     ).toBe('MOCK_ID_4');
   });
+
+  it('should not have required class and aria-required attribute and label should not have screen reader text when not required', () => {
+    componentInstance.labelText = 'Testing';
+    fixture.detectChanges();
+    const labelWrapper = getLabelEl();
+
+    expect(
+      labelWrapper?.classList.contains('sky-control-label-required'),
+    ).toBeFalse();
+    expect(labelWrapper?.querySelector('.sky-screen-reader-only')).toBeNull();
+  });
+
+  it('should have appropriate classes and label should have screen reader text when file is required', fakeAsync(() => {
+    componentInstance.labelText = 'Testing';
+    fixture.componentInstance.required = true;
+    fixture.detectChanges();
+    const labelWrapper = getLabelEl();
+
+    expect(
+      labelWrapper?.classList.contains('sky-control-label-required'),
+    ).toBeTrue();
+    expect(
+      labelWrapper?.querySelector('.sky-screen-reader-only')?.textContent,
+    ).toBe('Required');
+  }));
 
   it('should pass accessibility', async () => {
     fixture.detectChanges();