+++ libs/components/forms/testing/src/checkbox/checkbox-harness.spec.ts
@@ -155,6 +155,20 @@ describe('Checkbox harness', () => {
     );
   });
 
+  it('should get the hint text', async () => {
+    const { checkboxHarness, fixture } = await setupTest({
+      dataSkyId: 'my-phone-checkbox',
+    });
+    const hintText = 'Hint text for the checkbox.';
+
+    await expectAsync(checkboxHarness.getHintText()).toBeResolvedTo('');
+
+    fixture.componentInstance.phoneHintText = hintText;
+    fixture.detectChanges();
+
+    await expectAsync(checkboxHarness.getHintText()).toBeResolvedTo(hintText);
+  });
+
   it('should get the checkbox name and value', async () => {
     const { checkboxHarness } = await setupTest({
       dataSkyId: 'my-email-checkbox',
@@ -212,6 +226,53 @@ describe('Checkbox harness', () => {
 
     await expectAsync(checkboxHarness.hasRequiredError()).toBeRejectedWithError(
       'No form errors found.',
+    );
+  });
+
+  it('should throw an error if no help inline is found', async () => {
+    const { checkboxHarness } = await setupTest({
+      dataSkyId: 'my-mail-checkbox',
+    });
+
+    await expectAsync(checkboxHarness.clickHelpInline()).toBeRejectedWithError(
+      'No help inline found.',
+    );
+  });
+
+  it('should open help inline popover', async () => {
+    const { checkboxHarness, fixture } = await setupTest({
+      dataSkyId: 'my-phone-checkbox',
+    });
+    await checkboxHarness.clickHelpInline();
+    fixture.detectChanges();
+    fixture.whenStable();
+
+    await expectAsync(checkboxHarness.getHelpPopoverContent()).toBeResolved();
+  });
+
+  it('should get help popover content', async () => {
+    const { checkboxHarness, fixture } = await setupTest({
+      dataSkyId: 'my-phone-checkbox',
+    });
+    await checkboxHarness.clickHelpInline();
+    fixture.detectChanges();
+    fixture.whenStable();
+
+    await expectAsync(checkboxHarness.getHelpPopoverContent()).toBeResolvedTo(
+      '(xxx)xxx-xxxx',
+    );
+  });
+
+  it('should get help popover title', async () => {
+    const { checkboxHarness, fixture } = await setupTest({
+      dataSkyId: 'my-phone-checkbox',
+    });
+    await checkboxHarness.clickHelpInline();
+    fixture.detectChanges();
+    fixture.whenStable();
+
+    await expectAsync(checkboxHarness.getHelpPopoverTitle()).toBeResolvedTo(
+      'Format',
     );
   });
 });