+++ libs/components/forms/testing/src/checkbox/fixtures/checkbox-harness-test.component.ts
@@ -15,8 +15,12 @@ export class CheckboxHarnessTestComponent {
   public myForm: UntypedFormGroup;
   public hideEmailLabel = false;
   public mailControl: UntypedFormControl;
+  public helpPopoverContent: string | undefined;
+  public helpPopoverTitle: string | undefined;
   public hidePhoneLabel = false;
   public hideGroupLabel = false;
+  public hintText: string | undefined;
+  public phoneHintText: string | undefined;
 
   #formBuilder: UntypedFormBuilder;